requirements.txt
beautifulsoup4==4.4.1
bs4==0.0.1
Flask==0.11.1
Flask-Cors==2.1.2
html2text==2016.5.29
httplib2==0.9.2
ipaddress==1.0.16
Jinja2==2.8
JSON-log-formatter==0.1.0
mongoengine==0.10.6
nltk==3.2.1
scipy==0.17.1
numpy==1.11.0
pandas==0.18.1
parsedatetime==2.1
python-crfsuite==0.8.4
python-dateutil==2.5.3
requests==2.10.0
scikit-learn==0.17.1
simplejson==3.8.2
Werkzeug==0.11.10
attrs==15.2.0
cffi==1.6.0
cryptography==1.3.2
cssselect==0.9.1
lxml==3.6.0
parsel==1.0.2
pyasn1-modules==0.0.8
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
PyYAML==3.11
queuelib==1.4.2
requests==2.10.0
Scrapy==1.2.0
selenium==2.53.2
Twisted==16.2.0
zope.interface==4.1.3
translate
psycopg2==2.6.2
bs4==0.0.1
requests-cache==0.4.12
pybloom==1.1
